Refactored Plaid so public API methods are obvious
  - Marked public API methods, added some comments
  - Merged Plaid::AddUser into Plaid
  - Dropped `Plaid.existing_user` -- use `Plaid.set_user`
  - Added public API method that computes fully-qualified token based
    upon institution type. This is not documented in the Plaid API docs,
    so this needs investigation
  - Reorganized specs to better fit the code reorganization
