+++ spec/plaid_spec.rb
@@ -170,5 +170,28 @@ RSpec.describe Plaid do
     let(:access_token) { 'test' }
 
     it { expect(subject.access_token).to eq(access_token)}
+
+    context 'gets a valid user with accounts and transactions' do
+      let(:user) { Plaid.set_user('test_wells',['connect']) }
+      it { expect(user.transactions).not_to be_empty }
+    end
+
+    context 'gets a valid user with accounts' do
+      let(:user) { Plaid.set_user('test_wells',['auth']) }
+      it { expect(user.accounts).not_to be_empty }
+    end
+
+    #TODO: Fully vet the info api endpoint for the beta functions before adding this as a supported function.
+    pending 'need to vet the info api endpoint' do
+      context 'gets a valid user with info' do
+        let(:user) { Plaid.set_user('test_wells',['info']) }
+        it { expect(user.accounts).to be_truthy}
+      end
+
+      context 'gets a fully validated user with all access granted' do
+        let(:user) { Plaid.set_user('test_wells',['connect','info','auth']) }
+        it { expect(user.transactions).to be_truthy}
+      end
+    end
   end
 end
